requirements.txt
python-dotenv==1.0.0
yt-dlp>=2024.0.0
